Move some ICs into src
